requirements.txt
alpaca-py
scikit-learn
pandas
numpy
matplotlib
mlflow
streamlit
evidently
python-dotenv
pytest
ta
xgboost
